requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.3
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
jinja2==3.1.3
aiofiles==23.2.1
wsgidav==4.3.0
python-magic==0.4.27
PyPDF2==3.0.1
ebooklib==0.18
pillow==10.2.0 